+++ internal/environment/doc.go
@@ -1,0 +1,3 @@
+// Package environment is an abstraction that Ferrite uses to interact with the
+// environment.
+package environment