+++ internal/mode/config.go
@@ -9,11 +9,11 @@ import (
 
 // Config is the configuration used when running a mode.
 type Config struct {
-	Registry *variable.Registry
-	Args     []string
-	Out      io.Writer
-	Err      io.Writer
-	Exit     func(int)
+	Registries variable.RegistrySet
+	Args       []string
+	Out        io.Writer
+	Err        io.Writer
+	Exit       func(int)
 }
 
 // DefaultConfig is the default configuration for running a mode.
@@ -23,11 +23,10 @@ var DefaultConfig Config
 // largely intended for tearing down tests.
 func ResetDefaultConfig() {
 	DefaultConfig = Config{
-		&variable.DefaultRegistry,
-		os.Args,
-		os.Stdout,
-		os.Stderr,
-		os.Exit,
+		Args: os.Args,
+		Out:  os.Stdout,
+		Err:  os.Stderr,
+		Exit: os.Exit,
 	}
 }