+++ option_withregistry.go
@@ -1,9 +1,7 @@
 package ferrite
 
-import "github.com/dogmatiq/ferrite/internal/variable"
-
-// WithRegistry is an option that sets the variable registry to use.
-func WithRegistry(reg *variable.Registry) interface {
+// WithRegistry is an option that sets the variable registries to use.
+func WithRegistry(reg Registry) interface {
 	InitOption
 	RequiredOption
 	OptionalOption
@@ -15,10 +13,10 @@ func WithRegistry(reg *variable.Registry) interface {
 
 	return option{
 		ApplyToInitConfig: func(cfg *initConfig) {
-			cfg.ModeConfig.Registry = reg
+			cfg.ModeConfig.Registries.Add(reg)
 		},
 		ApplyToSetConfig: func(cfg *variableSetConfig) {
-			cfg.Registry = reg
+			cfg.Registries = append(cfg.Registries, reg)
 		},
 	}
 }