+++ varset_deprecated.go
@@ -43,7 +43,7 @@ func deprecated[T any, Schema variable.TypedSchema[T]](
 	}
 
 	v := variable.Register(
-		cfg.Registry,
+		cfg.Registries,
 		b.Done(s),
 	)